+++ doc/WhatsNew.rst
@@ -24,6 +24,7 @@ Ver 4.1.0 (2022-06-30)
   file in .../ginga/examples/configs)
 - Fixed an issue with readout of values under the cursor when an image
   is plotted at a non-zero origin
+- Add Contrast and Brightness adjustments in "Preferences" plugin
 
 Ver 4.0.1 (2022-12-27)
 ======================